Правка всех тестов аппки и объединение файлов data
