+++ global-setup.ts
@@ -1,5 +1,5 @@
 import { chromium } from '@playwright/test'
-import { Credentials} from './src/data/login'
+import { Credentials} from './src/data/appData'
 
 
 const globalSetup = async (): Promise<void> => {
