--- src/data/aboutModal.ts
@@ -1,3 +1,0 @@
-export enum AboutData {
-    About_Us = 'About us'
-}
