+++ src/data/appData.ts
@@ -1,0 +1,88 @@
+// Login Page
+export const Credentials = {
+    CorrectUsername:"Test2022",
+    CorrectPassword:"Test2022",
+    NotCorrectUsername:"454545qq",
+    NotCorrectPassword:"454545qq",
+    EmptyUsername:'',
+    EmptyPassword:''
+}
+
+export const VerificationText = {
+    LogIn:"Log in",
+    Name: "Welcome Test2022",
+    Style: "style",
+    DisplayNone:"display: none;",
+    DisplayBlock:"display: block;",
+    Class:"class",
+    ModalFadeShow:"modal fade show",
+    ModalFade:"modal fade",
+}
+
+export const ErrorsText = {
+    UserNotExist:"User does not exist.",
+    WrongPassword:"Wrong password.",
+    ValidationError:"Please fill out Username and Password.",
+}
+
+
+// Home page
+
+export const Headers = {
+    PRODUCT_STORE: "PRODUCT STORE",
+    Home: "Home",
+    Contact: "Contact",
+    AboutUs: "About us",
+    Cart: "Cart",
+    Logout: "Log out",
+    Name:"Welcome Test2022",
+ }
+ export const Categories = {
+    Categories: "CATEGORIES",
+    Phones: "Phones",
+    Laptops: "Laptops",
+    Monitors:"Monitors",
+    Alt: "alt",
+    FirstSlide: "First slide",
+    SecondSlide: "Second slide",
+    ThirdSlide: "Third slide",
+ }
+ 
+
+ //Contact Modal
+
+ export enum DataString {
+    Title = 'New message',
+    Email = 'sha@mail.ru',
+    Name = 'Mike',
+    Message = 'My name is Mike. Hello guys'
+}
+
+// About Modal
+
+export enum AboutData {
+    About_Us = 'About us'
+}
+
+//Item
+
+export const Titles = {
+    samsungGalaxySix: 'Samsung galaxy s6',
+    sonyXperiaZFive: 'Sony xperia z5',
+    macBookAir: 'MacBook air',
+    macBookPro: 'MacBook Pro',
+}
+
+export const Price = {
+    samsungGalaxySix: '$360 *includes tax',
+    sonyXperiaZFive: '$320 *includes tax',
+    macBookAir: '$700 *includes tax',
+    macBookPro: '$1100 *includes tax',
+}
+
+export const MoreInformation = {
+    samsungGalaxySix: 'Product descriptionThe Samsung Galaxy S6 is powered by 1.5GHz octa-core Samsung Exynos 7420 processor and it comes with 3GB of RAM. The phone packs 32GB of internal storage cannot be expanded.',
+    sonyXperiaZFive: 'Product descriptionSony Xperia Z5 Dual smartphone was launched in September 2015. The phone comes with a 5.20-inch touchscreen display with a resolution of 1080 pixels by 1920 pixels at a PPI of 424 pixels per inch.',
+    macBookAir: 'Product description1.6GHz dual-core Intel Core i5 (Turbo Boost up to 2.7GHz) with 3MB shared L3 cache Configurable to 2.2GHz dual-core Intel Core i7 (Turbo Boost up to 3.2GHz) with 4MB shared L3 cache.',
+    macBookPro: "Product descriptionApple has introduced three new versions of its MacBook Pro line, including a 13-inch and 15-inch model with the Touch Bar, a thin, multi-touch strip display that sits above the MacBook Pro's keyboard."
+}
