--- src/data/contactModal.ts
@@ -1,6 +1,0 @@
-export enum DataString {
-    Title = 'New message',
-    Email = 'sha@mail.ru',
-    Name = 'Mike',
-    Message = 'My name is Mike. Hello guys'
-}
