--- src/data/home.ts
@@ -1,21 +1,0 @@
-export const Headers = {
-   PRODUCT_STORE: "PRODUCT STORE",
-   Home: "Home",
-   Contact: "Contact",
-   AboutUs: "About us",
-   Cart: "Cart",
-   Logout: "Log out",
-   Name:"Welcome Test2022",
-}
-export const Categories = {
-   Categories: "CATEGORIES",
-   Phones: "Phones",
-   Laptops: "Laptops",
-   Monitors:"Monitors",
-   Alt: "alt",
-   FirstSlide: "First slide",
-   SecondSlide: "Second slide",
-   ThirdSlide: "Third slide",
-}
-
-
