--- src/data/item.ts
@@ -1,20 +1,0 @@
-export const Titles = {
-    samsungGalaxySix: 'Samsung galaxy s6',
-    sonyXperiaZFive: 'Sony xperia z5',
-    macBookAir: 'MacBook air',
-    macBookPro: 'MacBook Pro',
-}
-
-export const Price = {
-    samsungGalaxySix: '$360 *includes tax',
-    sonyXperiaZFive: '$320 *includes tax',
-    macBookAir: '$700 *includes tax',
-    macBookPro: '$1100 *includes tax',
-}
-
-export const MoreInformation = {
-    samsungGalaxySix: 'Product descriptionThe Samsung Galaxy S6 is powered by 1.5GHz octa-core Samsung Exynos 7420 processor and it comes with 3GB of RAM. The phone packs 32GB of internal storage cannot be expanded.',
-    sonyXperiaZFive: 'Product descriptionSony Xperia Z5 Dual smartphone was launched in September 2015. The phone comes with a 5.20-inch touchscreen display with a resolution of 1080 pixels by 1920 pixels at a PPI of 424 pixels per inch.',
-    macBookAir: 'Product description1.6GHz dual-core Intel Core i5 (Turbo Boost up to 2.7GHz) with 3MB shared L3 cache Configurable to 2.2GHz dual-core Intel Core i7 (Turbo Boost up to 3.2GHz) with 4MB shared L3 cache.',
-    macBookPro: "Product descriptionApple has introduced three new versions of its MacBook Pro line, including a 13-inch and 15-inch model with the Touch Bar, a thin, multi-touch strip display that sits above the MacBook Pro's keyboard."
-}
