--- src/data/login.ts
@@ -1,25 +1,0 @@
-export const Credentials = {
-    CorrectUsername:"Test2022",
-    CorrectPassword:"Test2022",
-    NotCorrectUsername:"454545qq",
-    NotCorrectPassword:"454545qq",
-    EmptyUsername:'',
-    EmptyPassword:''
-}
-
-export const VerificationText = {
-    LogIn:"Log in",
-    Name: "Welcome Test2022",
-    Style: "style",
-    DisplayNone:"display: none;",
-    DisplayBlock:"display: block;",
-    Class:"class",
-    ModalFadeShow:"modal fade show",
-    ModalFade:"modal fade",
-}
-
-export const ErrorsText = {
-    UserNotExist:"User does not exist.",
-    WrongPassword:"Wrong password.",
-    ValidationError:"Please fill out Username and Password.",
-}
