+++ src/tests/aboutModal.spec.ts
@@ -1,5 +1,5 @@
 import {test, expect} from "@playwright/test"
-import { AboutData } from "../data/aboutModal"
+import { AboutData } from "../data/appData"
 import { AboutModal } from "../pages/aboutModal"
 
 
@@ -11,19 +11,21 @@ test.beforeEach(async({page})=>{
     aboutModal = new AboutModal(page)
     await page.goto('https://www.demoblaze.com/index.html')
     await aboutModal.aboutUsButtonHeader.click()
-    await aboutModal.loadPage()
-  
+    await aboutModal.loadPage() 
 })
 
-test.describe('Общие проверки', async()=>{
+test.describe('Общие проверки модального окна About Us', async()=>{
+
+    test('Нажать на About Us -> открывается модальное окно', async()=>{
+        await expect(aboutModal.videoModal).toHaveAttribute('class', 'modal fade show')
+     })
 
     test('Элементы модального окна', async()=>{
         await expect(aboutModal.videoModalLabel).toBeVisible()
         await expect(aboutModal.videoModalLabel).toHaveText(AboutData.About_Us)
         await expect(aboutModal.closeModalCross).toBeVisible()
         await expect(aboutModal.videoBlock).toBeVisible()
-        await expect(aboutModal.closeModalButton).toBeVisible()
-        
+        await expect(aboutModal.closeModalButton).toBeVisible()        
     })
 
     test('Нажать на кнопку крестик -> модальное окно контактов закрывается', async()=>{
@@ -37,3 +39,5 @@ test.describe('Общие проверки', async()=>{
     })
 })
 
+
+// PWDEBUG=1 npx playwright test
