+++ src/tests/cart.spec.ts
@@ -49,7 +49,7 @@ test.describe('Общие проверки страницы карточки', a
 })
 
 
-test.describe('Добавление разных продуктов на страницу карточки', async()=>{
+test.describe('Добавление продуктов на страницу Cart', async()=>{
 
 test.describe('Старница по умолчанию', async()=>{
